Move source code for the synchronizer.
