+++ galley/pkg/kube/kube.go
@@ -12,7 +12,7 @@
 //  See the License for the specific language governing permissions and
 //  limitations under the License.
 
-package common
+package kube
 
 import (
 	"k8s.io/apiextensions-apiserver/pkg/client/clientset/clientset"
