+++ galley/pkg/kube/sync/common/const.go
