+++ galley/pkg/kube/sync/common/equals.go
