+++ galley/pkg/kube/sync/common/equals_test.go
