+++ galley/pkg/kube/sync/crd/equals.go
