+++ galley/pkg/kube/sync/crd/equals_test.go
