+++ galley/pkg/kube/sync/crd/operations.go
