+++ galley/pkg/kube/sync/crd/operations_test.go
