+++ galley/pkg/kube/sync/crd/rewrite.go
