+++ galley/pkg/kube/sync/crd/rewrite_test.go
