+++ galley/pkg/kube/sync/crd/synchronizer_test.go
