+++ galley/pkg/kube/sync/mapping.go
@@ -12,7 +12,7 @@
 //  See the License for the specific language governing permissions and
 //  limitations under the License.
 
-package server
+package sync
 
 import (
 	"k8s.io/apimachinery/pkg/runtime/schema"
