+++ galley/pkg/kube/sync/resource/equals.go
