+++ galley/pkg/kube/sync/resource/equals_test.go
