+++ galley/pkg/kube/sync/resource/operations.go
