+++ galley/pkg/kube/sync/resource/operations_test.go
