+++ galley/pkg/kube/sync/resource/rewrite.go
