+++ galley/pkg/kube/sync/resource/rewrite_test.go
