+++ galley/pkg/kube/sync/resource/synchronizer.go
