+++ galley/pkg/kube/sync/resource/synchronizer_test.go
